Fix this crash that would sometimes occur on exit:
>	FoenixIDE.exe!FoenixIDE.UI.MainWindow.Write_CPS_FPS_Safe(string CPS, string FPS) Line 686	C#
 	FoenixIDE.exe!FoenixIDE.UI.MainWindow.Gpu_Update_Cps_Fps() Line 717	C#
 	FoenixIDE.exe!FoenixIDE.Display.Gpu.GpuRefreshTimer_Tick(object sender, System.EventArgs e) Line 87	C#
 	FoenixIDE.exe!KGySoft.CoreLibraries.HiResTimer.ExecuteTimer() Line 217	C#

Root cause of crash:

Sometimes, the following happens
- User closes the application
- MainWindow_FormClosing is called
- MainWindow_FormClosing starts tearing stuff down, including the gpu object, which owns the hires timer. Hires timer has a worker thread
- MainWindow_FormClosing prepares the hires timer to stop, by setting a bool that the timer loops on.
- But the thread is still running. It'll still finish the iteration of the loop it's currently on.

- MainWindow_FormClosing moves on and starts tearing other stuff down, including the main application thread.
- Since the timer thread is still running, it reaches this line
        Elapsed?.Invoke(this, new HiResTimerElapsedEventArgs(delay, fallouts));
which crashes because since the application thread has already exited.

Repro about 10% of the time. you have to get (un)lucky and have the timer thread be in the right spot for it to hit.

Fix: have MainWindow_FormClosing actually stop the timer thread, don't let it continue the current iteration. It looks like this was the intention. Rename the function in gpu from 'StopTimer' to 'DeleteTimer' to reflect this.