Fix a hang that can happen on slower execution environments, when there is mouse movement while loading FMX and possibly other kernels.
The problematic code is

           ; buffer some mouse movement immediately before
           PHA
wait    LDA @1 KDCTL
           BIT #2
           BNE wait

Before, the 'wait' loop would loop forever.
Reason: reads of KDCTL would not progress the emulated keyboard+mouse controller. Only reads of KDATA would.

In reality for the hardware, no CPU operation should really be dictating progress of the controller. The controller (8042-based, SuperIO) progresses on its own separate timeline. If we observe a need for more accuracy from this emulator, we should change the emulator to replicate that. For now, it's good enough to progress the controller on both kinds of reads that poll for it (reads of KDATA, and KDCTL), rather than just one of them. I observed this fixes the class of hang in the above kernelcode.