Fixed the interrupt checkboxes and the tooltips in the CPU Window to match the F256 or C256 registers.
Fixed an issue when loading .bin files - this now goes only to RAM, not devices.
Re-adjusted controls position in the Main and CPU windows and locked the forms.
Fixed issue when the machine was left in Jr(816) it would restart in C mode.
Changing the board version in the Main Window now is properly reflected in the CPU Window.

# Conflicts:
#	Main/FoenixSystem.cs
#	Main/UI/CPUWindow.Designer.cs
#	Main/UI/MainWindow.Designer.cs
#	Main/UI/MainWindow.cs